fix: Make tldraw work every time
